add user to all methods

+++ src/main/java/com/example/simple_todo_app/services/TaskServiceImpl.java
@@ -1,14 +1,15 @@
 package com.example.simple_todo_app.services;
 
-import com.example.simple_todo_app.exceptions.EmptyListException;
-import com.example.simple_todo_app.exceptions.MissingDataException;
-import com.example.simple_todo_app.exceptions.NotFoundException;
-import com.example.simple_todo_app.exceptions.OverExtendedLengthException;
+import com.example.simple_todo_app.exceptions.*;
 import com.example.simple_todo_app.models.Task;
+import com.example.simple_todo_app.models.User;
 import com.example.simple_todo_app.models.dtos.CreateNewTaskDTO;
 import com.example.simple_todo_app.models.dtos.TaskDTO;
 import com.example.simple_todo_app.repositories.TaskRepository;
+import com.example.simple_todo_app.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,62 +18,76 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class TaskServiceImpl implements TaskService {
+public class TaskServiceImpl {
 
     private final TaskRepository taskRepository;
+    private final UserRepository userRepository;
 
-    @Override
     public Task createTask(CreateNewTaskDTO taskDTO) {
 
         if (taskDTO.getTitle() == null || taskDTO.getTitle().isEmpty()) {
             throw new MissingDataException("Task title");
         }
+
         if (taskDTO.getTitle().length() > 50) {
             throw new OverExtendedLengthException("Title");
         }
+
+        User user = getAuthenticatedUser();
+
         Task task = Task.builder()
                 .title(taskDTO.getTitle())
+                .user(user)
                 .completed(false)
                 .build();
+
         taskRepository.save(task);
         return task;
     }
 
-    @Override
     public List<TaskDTO> getAllTasksByStatus(String status) {
         List<Task> tasks;
 
+        User user = getAuthenticatedUser();
+
         if ("completed".equals(status)) {
-            tasks = taskRepository.findAllByCompletedTrue();
+            tasks = taskRepository.findAllByUserAndCompletedTrue(user);
+
         } else if ("active".equals(status)) {
-            tasks = taskRepository.findAllByCompletedFalse();
+            tasks = taskRepository.findAllByUserAndCompletedFalse(user);
+
         } else {
-            tasks = taskRepository.findAll();
+            tasks = taskRepository.findAllByUser(user);
         }
 
         return tasks.stream().map(TaskDTO::new).toList();
     }
 
-    @Override
     @Transactional
     public void deleteAllCompletedTasks() {
-        if (taskRepository.findAllByCompletedTrue().isEmpty()) {
+
+        User user = getAuthenticatedUser();
+
+        if (taskRepository.findAllByUserAndCompletedTrue(user).isEmpty()) {
             throw new EmptyListException("Completed tasks");
         }
-        taskRepository.deleteAllByCompletedTrue();
+        taskRepository.deleteAllByUserAndCompletedTrue(user);
     }
 
-    @Override
     public void deleteById(Long id) {
         Optional<Task> task = taskRepository.findById(id);
         if (task.isEmpty()) {
             throw new NotFoundException("Task with id " + id);
         }
         Task taskToDelete = task.get();
+        User user = getAuthenticatedUser();
+        if (!taskToDelete.getUser().equals(user)) {
+            throw new UnauthorizedException();
+        }
+
         taskRepository.delete(taskToDelete);
     }
 
-    @Override
     public Task updateTitleById(Long id, String newTitle) {
         Optional<Task> task = taskRepository.findById(id);
         if (task.isEmpty()) {
@@ -85,20 +100,36 @@ public class TaskServiceImpl implements TaskService {
             throw new OverExtendedLengthException("Title");
         }
         Task taskToUpdate = task.get();
+        User user = getAuthenticatedUser();
+        if (!taskToUpdate.getUser().equals(user)) {
+            throw new UnauthorizedException();
+        }
         taskToUpdate.setTitle(newTitle);
         taskRepository.save(taskToUpdate);
         return taskToUpdate;
     }
 
-    @Override
     public Task updateCompletedById(Long id) {
         Optional<Task> task = taskRepository.findById(id);
         if (task.isEmpty()) {
             throw new NotFoundException("Task with id " + id);
         }
         Task taskToUpdate = task.get();
+        User user = getAuthenticatedUser();
+        if (!taskToUpdate.getUser().equals(user)) {
+            throw new UnauthorizedException();
+        }
         taskToUpdate.setCompleted(!taskToUpdate.getCompleted());
         taskRepository.save(taskToUpdate);
         return taskToUpdate;
+    }
+
+    public User getAuthenticatedUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        String userEmail = authentication.getName();
+
+        return userRepository.findByEmail(userEmail)
+                .orElseThrow(() -> new NotFoundException("User with email " + userEmail));
     }
 }
